backend/requirements.txt
fastapi
uvicorn
openai>=1.40.0
supabase
pydantic
python-dotenv
